src/main/java/seo/dale/practice/java/util/predicate/FilteringApples.java
package seo.dale.practice.java.util.predicate;

import java.util.ArrayList;
import java.util.List;
import java.util.function.Predicate;

/**
 * https://github.com/java8/Java8InAction/blob/master/src/main/java/lambdasinaction/chap1/FilteringApples.java
 */
public class FilteringApples {

	public static List<Apple> filterApples(List<Apple> inventory, Predicate<Apple> p){
		List<Apple> result = new ArrayList<>();
		for(Apple apple : inventory){
			if(p.test(apple)){
				result.add(apple);
			}
		}
		return result;
	}

	public static boolean isGreenApple(Apple apple) {
		return "green".equals(apple.getColor());
	}

	public static boolean isHeavyApple(Apple apple) {
		return apple.getWeight() > 150;
	}

}
